core_app/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.6.0
django-storages==1.11
djangorestframework==3.12.2
Pillow==8.0.1
psycopg2-binary==2.8.6
python-decouple==3.3
pytz==2020.4
sqlparse==0.4.1
